+++ jkube-kit/resource/helm/src/test/java/org/eclipse/jkube/kit/resource/helm/HelmDependencyTest.java
@@ -28,11 +28,15 @@ class HelmDependencyTest {
         .name("name")
         .repository("repository")
         .version("version")
+        .alias("alias")
+        .condition("is.enabled")
         .build();
 
     HelmDependency sameHelmDependency = HelmDependency
         .builder()
         .name("name")
+        .alias("alias")
+        .condition("is.enabled")
         .repository("repository")
         .version("version")
         .build();
@@ -59,6 +63,7 @@ class HelmDependencyTest {
     assertThat(helmDependency)
         .hasFieldOrPropertyWithValue("name", "name")
         .hasFieldOrPropertyWithValue("repository", "repository")
-        .hasFieldOrPropertyWithValue("version", "version");
+        .hasFieldOrPropertyWithValue("version", "version")
+        .hasNoNullFieldsOrPropertiesExcept("alias", "condition");
   }
 }